feat: 支持中文GB2312字体，并可以显示中文字符

+++ TomoyoAfterHook/hook.h
@@ -33,7 +33,7 @@ typedef DWORD(__fastcall* ReadSeenDataFuncPtr)(BYTE ctx[64], SeenData* out, DWOR
 void HookInit();
 void HookDestroy();
 void RunDump();
-BYTE* GetSeenPatchData(int num);
+BYTE* GetSeenPatchData(size_t num);
 
 #endif
